chore: 패키지명 ìcoffeemanager로 변경 및 정리
