+++ src/main/java/coffeemanager/IndexController.java
@@ -1,0 +1,15 @@
+package coffeemanager;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/")
+public class IndexController {
+    @GetMapping
+    public String index() {
+        return "index";
+    }
+}
