+++ src/main/java/coffeemanager/app/controller/api/coffee/CoffeeApiController.java
@@ -1,0 +1,5 @@
+package coffeemanager.app.controller.api.coffee;
+
+public class CoffeeApiController {
+
+}
