+++ src/main/java/coffeemanager/app/controller/web/coffee/CoffeeController.java
@@ -1,0 +1,5 @@
+package coffeemanager.app.controller.web.coffee;
+
+public class CoffeeController {
+
+}
