+++ src/main/java/coffeemanager/app/model/auth/code/Role.java
@@ -1,0 +1,10 @@
+package coffeemanager.app.model.auth.code;
+
+import org.apache.ibatis.type.Alias;
+
+@Alias("MEMBER_ROLE")
+public enum Role {
+    ANONYMOUS,
+    ROLE_USER,
+    ROLE_ADMIN
+}
