+++ src/main/java/coffeemanager/app/model/coffee/CoffeeRepository.java
@@ -1,0 +1,6 @@
+package coffeemanager.app.model.coffee;
+
+public interface CoffeeRepository
+{
+
+}
