+++ src/main/java/coffeemanager/app/model/coffee/CoffeeService.java
@@ -1,0 +1,5 @@
+package coffeemanager.app.model.coffee;
+
+public class CoffeeService {
+
+}
