+++ src/main/java/coffeemanager/infra/response/ResponseCode.java
@@ -1,0 +1,32 @@
+package coffeemanager.infra.response;
+
+import org.springframework.http.HttpStatus;
+
+public enum ResponseCode {
+    OK("0000", HttpStatus.OK, "정상적으로 완료되었습니다."),
+    BAD_REQUEST("4000", HttpStatus.BAD_REQUEST, "잘못된 요청입니다."),
+    UNAUTHORIZED("4001", HttpStatus.UNAUTHORIZED, "권한이 없습니다."),
+    INTERNAL_SERVER_ERROR("5000", HttpStatus.INTERNAL_SERVER_ERROR, "서버에러 입니다.");
+    
+    private final String code;
+    private final HttpStatus status;
+    private final String message;
+    
+    ResponseCode(String code, HttpStatus status, String message) {
+        this.code = code;
+        this.status = status;
+        this.message = message;
+    }
+    
+    public String code() {
+        return code;
+    }
+    
+    public HttpStatus status() {
+        return status;
+    }
+    
+    public String message() {
+        return message;
+    }
+}
