--- src/main/java/com/grepp/spring/IndexController.java
@@ -1,15 +1,0 @@
-package com.grepp.spring;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/")
-public class IndexController {
-    @GetMapping
-    public String index() {
-        return "index";
-    }
-}
