--- src/main/java/com/grepp/spring/app/controller/api/coffee/CoffeeApiController.java
@@ -1,5 +1,0 @@
-package com.grepp.spring.app.controller.api.coffee;
-
-public class CoffeeApiController {
-
-}
