--- src/main/java/com/grepp/spring/app/controller/web/coffee/CoffeeController.java
@@ -1,5 +1,0 @@
-package com.grepp.spring.app.controller.web.coffee;
-
-public class CoffeeController {
-
-}
