--- src/main/java/com/grepp/spring/app/model/auth/domain/Principal.java
@@ -1,22 +1,0 @@
-package com.grepp.spring.app.model.auth.domain;
-
-import com.grepp.spring.app.model.member.dto.Member;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-
-public class Principal extends User {
-    
-    public Principal(String username, String password,
-        Collection<? extends GrantedAuthority> authorities) {
-        super(username, password, authorities);
-    }
-    
-    public static Principal createPrincipal(Member member,
-        List<SimpleGrantedAuthority> authorities){
-        return new Principal(member.getUserId(), member.getPassword(), authorities);
-    }
-}
