--- src/main/java/com/grepp/spring/app/model/coffee/CoffeeRepository.java
@@ -1,6 +1,0 @@
-package com.grepp.spring.app.model.coffee;
-
-public interface CoffeeRepository
-{
-
-}
