--- src/main/java/com/grepp/spring/app/model/coffee/CoffeeService.java
@@ -1,5 +1,0 @@
-package com.grepp.spring.app.model.coffee;
-
-public class CoffeeService {
-
-}
