+++ src/main/resources/mybatis/mappers/member/memberMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.grepp.spring.app.model.member.MemberRepository">
+<mapper namespace="coffeemanager.app.model.member.MemberRepository">
 
   <resultMap id="memberMap" type="Member">
     <id property="userId" column="user_id"/>
